+++ interactions/btn/randomBtn.js
@@ -1,6 +1,5 @@
-const {civilWarWindow} = require("../../components/gameIntroWindow.js");
 module.exports = {
-	name:"startBtn",
+	name:"randomBtn",
 	async execute(interaction) {
 		const client=interaction.client;
         const waitingRoom=client.waitingRooms.get(interaction.guildId);
@@ -9,9 +8,16 @@ module.exports = {
             interaction.reply("Only hosts can interact!");
             return;
         }
-		
-		waitingRoom.MoveEachTeamToVoiceChannel();
-		waitingRoom.SendOpeningWindow();
+
+		if (!waitingRoom.CheckPlayersEven()){
+			interaction.reply("Players are not even!");
+            return;
+		}
+
+		waitingRoom.SetStageRandom();
+        waitingRoom.MakeTeam();
+		waitingRoom.SetTeam();
+        waitingRoom.SendStageWindow();
 		interaction.message.delete();
 		return;
 	},
